net: ethernet: ti: fix netdevice stats for XDP
Align netdevice statistics when the device is running in XDP mode
to other upstream drivers. In particular reports to user-space rx
packets even if they are not forwarded to the networking stack
(XDP_PASS) but they are redirected (XDP_REDIRECT), dropped (XDP_DROP)
or sent back using the same interface (XDP_TX). This patch allows the
system administrator to very the device is receiving data correctly.